+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG for Sulu Sales
 ========================
 
+* 0.4.2 (2016-07-28)
+
+    * BUGFIX      [General]       Set parent item if an addon is added.
+    * BUGFIX      [CoreBundle]    Fixed relation between item and parent item.
+
 * 0.4.1 (2016-07-19)
 
     * FEATURE     [General]       Updated for usage with recurring price calculation